pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-serverless"
version = "1.0.0"
description = "Simplified serverless Discord bot interactions"
readme = "README.md"
authors = [{ name = "Ross Bamford", email = "roscopeco@gmail.com" }]
license = { file = "LICENSE.md" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["feed", "reader", "tutorial"]
dependencies = [
    "pynacl >= 1.5.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["black", "bumpver", "pre-commit", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/roscopeco/discord-serverless"

[tool.bumpver]
current_version = "2023.1001-alpha"
version_pattern = "YYYY.BUILD[-TAG]"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]
"README.md" = [
    "{version}",
    "{pep440_version}",
]

[tool.black]
py36 = true
include = '\.pyi?$'
exclude = '''
    /(
        \.git
      | \.venv
      | _build
      | build
      | dist
    )/
    '''
